pyproject.toml
[tool.poetry]
name = "cdk-gw-tools"
description = "Conduktor Proxy API Client"
version = "0.3.1"
authors = ["John \"Preston\" Mille <john@ews-network.net>"]
readme = "README.md"
license = "LICENSE"
packages = [{ include = "cdk_gw_tools" }]
keywords = ["compose-x", "conduktor", "kafka", "proxy"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

include = [
    "cdk_gw_tools/**/*.json",
]

exclude = ["*.pyc", "*~", "*pycache*"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/JohnPreston/cdk-gw-tools/issues"
"Source (GitHub)" = "https://github.com/JohnPreston/cdk-gw-tools"
"Source (CodeBerg)" = "https://codeberg.org/JohnPreston/cdk-gw-tools"

[tool.poetry.dependencies]
python = "^3.8"
cdk-proxy-api-client = "^0.7"
compose-x-common = "^1.4.1"
jsonschema = ">=4.15,<4.18"
importlib-resources = "^6.1.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.1.1"
black = "^23.9"
isort = "^5.12"
pyupgrade = "^3.3.1"
tbump = "^6.9.0"
datamodel-code-generator = "^0.25.1"

[tool.poetry.scripts]
cdk-cli = "cdk_gw_tools.cli:main"


[tool.tbump]
github_url = "https://github.com/JohnPreston/cdk-gw-tools"

[tool.tbump.version]
current = "0.3.1"
regex = '''
  (?P<major>\d+)
  \.
  (?P<minor>\d+)
  \.
  (?P<patch>\d+)
  (?:(?P<rc>[\S]+))?
  '''

[tool.tbump.git]
message_template = "Bump to {new_version}"
tag_template = "v{new_version}"

[[tool.tbump.file]]
src = "pyproject.toml"

[[tool.tbump.file]]
src = "cdk_gw_tools/__init__.py"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = "kelvin"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
